Updated how CommuinicationEntities are created and handled
This commit features the conversion of "CommunicationEntityBar" to a
container component, CommunicationEntityBar will hold
CommunicationEntity components and will handle the logic behind which
CommunicationEntity has been selected by the user. This was done to
ensure that there is low coupling between the CommunicationEntity
components and SideBar, and the extra layer of abstraction (by
converting CommunicationEntityBar into a container-like component) will
also lead to higher cohesion.