Updated how actions for CommEntity are shown
This commit involves delegating the task of determining which CommEntity
should be shown to the CommunicationEntityBar component. This is
achieved by passing the down the handleContextMenu function from
CommunicationEntityBar down to its child CommunicationEntity components.
The child components will then call this via its onContextMenu prop to
indicate to the CommunicationEntityBar component which CommEntity should
be showing the actions.