Performance updates to search functionality
The search functionality raised some issues with performance and would
often cause the application to hang or crash on the frontend. As a fix
(might be permanent), the emitting of events that triggers a search query to
the database has been limited to only when the user presses the enter
key while the SearchBar component is in focus. That way, the displayed
CommunicationEntities are only updated when the user presses the enter
key, instead of every key press (not ideal but it will have to do for
now). Furthermore, SearchBar has been converted into a class based react
component temporarily to determine if there are any performance
differences between implemented SearchBar using react hooks or classes.
This is a temporary change for experimental purposes only.