Updated searching concept of application
Because there was a problem with determining what kind of actions to
place on CommunicationEntities, the searching functionality of the
application has been updated so that there is a clear separation between
searching for known communication entities (users the user are friends
with, or groups that the user wants to join) and searching for unkown
communication entities. An extra option was introduced in the
ToggleSwitch component in SideBar that will allow the user to search for
unknown CommunicationEntities, whilst the Friends and Group options
should only search for Friends or Groups already associated with the
user respectively. Note that all 3 options still retain the same
SearchBar component which is important for reusability purposes. The
SearchBar was converted into class based react component since the event
listener on its input was not correctly determining which socket.io
event to fire based on its mode props passed to it by the SideBar
component.