Updated database querying operations to use mongoose models
While the original approach of querying the database using native
mongodb querying (through DatabaseConnection instance) provides a good
generic platform to query the database, using DatabaseConnection to
query the database all the time can make the code more difficult to
understand and may make it more susceptible to long chains of callback
functions to achieve a simple query. Using the mongoose models to
perform the queries is a good compromise since it makes the code more
readable and easier to work with, but we lose the generic flexibility
that was offered with the original approach, and will narrow the scope of
the querying methods to specific collections. This tradeoff is worth it
since the database was never meant to work with no more than a few
collections at most. If the databse grows with more collections in the
future, we can refactor back to the original approach for better
scalability.