Updated how notifications are created
This commit updates the creation of notifications by strictly confining
their creation to the front end. This enables the code to achieve higher
cohesion and lower coupling since notification objects are no longer
being sent from the server to the client. All it takes now to create a
notification is for the server to fire certain events which the client
will react to by creating a notification in the front end only.