Fixed issue with receiving ADD_FRIEND_DENIED event multiple times in front end
The issue arose due to the fact that a listener for the event was being
added to the client's socket everytime SearchResultsBar was rendered. As
a solution, the useEffect hook was used with an empty array as the
second parameter so that the listeners for the event was only added
during the first render of SearchResultsBar.