Updated SearchBar functionality in ProfileView
The initial approach included the usage of 2 SearchBar components in
ProfileView, one dedicated to searching for thisUser's friends, the
other for searching group thisUser is involved with. This approach shed
light on the problem that there needs to be better isolation between
distinct instances of the same component. When one SearchBar component
was being interacted with, the results are also propagated to the other
SearchBar instance. The root of the problem is still unknown. As a fix,
ProfileView now only features 1 SearchBar instance and the user is given
the option to toggle between searching for groups or searching for their
friends. This ultimately leads to a cleaner UX and simplifies the code
of the component even more.