Updated how sockets are used in the application
This commit places the socket server setup in a different file and
allows io to be imported across the different files in the serverside.
This commit also establishes the pattern that in order for a socket
event to be triggered by the server, the clientside must first send the
socket id of the client in the body of the request sent via axios. This
socket id is then used to target the correct client to fire the socket
io event to.