Updated login mechanism
This ccommit changes the passage of thisUser object fromt he server to
the client by sending it as a response to the post request issued by the
client, instead of being passed with the triggering of a socket io
event. This commit also removes the creation of a notification that
indicates to the user that their login was succesful.