Began changing the meaning of View components
This commit features the refactoring of View components into ViewContent
components, whilst the meaning of the original View components was
changed into a 'containing component'. This refactor was necessary since
the different views needed to be represented by their own unique
react-router routes to easily enable express to setup proper routes for
the http requests. This was not possible in the previous configuration
since all the View components were represented by a single
react-router route (/chat) with each view being conditionally rendered
in this one route.