Used localStorage for thisUser
This was done so that tehe contents of the thisUser object would
persist even after a page refresh. It also makes it easier to
access thisUser across all the components as we no longer need
the long chain of passing thisUser props down to the deepest
child component just so that the child component can read the
contents of thisUser. This commit also added the removal of
certain SEARCH_EVENTS in the SearchBar component to prevent
no-op operations that occur in components that use the SearchBar
component.