Moved selectedNavOptionIndex in SideBar to sessionStorage
This was necessary since SideBar was re-rendered as a new
instance everytime a navOption in the SideBar was selected which
leads the user to a new react-route. Storing the state of
selectedNavOptionIndex in SideBar will make it difficult to
mutate this index appropriately everytime a navOption was
selected by the user.