Made express server and socket.io server use the same http server
Prior to this commit, the express server and socket.io server
used different http servers and were setup independent of each
other. This commit makes both servers use the same http server
on the same port number. As a result there were some changes
made with the way socket operations are triggerred within the
express-related routes. There were problems with importing io
from server.js so therefore, the routes no longer use sockets to
trigger notifications in the client side. Instead, the server
will reply to the request of the user through res.send using a
boolean to indicate whether or not the request made by the
client was successful or not. This boolean is then checked in
the client side to show the correct notification.