Overhauled UserInfoForm component
This overhaul was primarily due to the fact that user passwords are now
hashed to protect them. The original design of UserInfoForm involved
fetching the password in raw form from the database for display in a
password type input field. Now that the passwords are hashed, this is no
longer a practical design solution. This overhaul removes the need of
the user to enter their password to update their account information.
Instead, they will be allowed to update their information as soon as the
component mounts, but to save changes, the user has to enter their
password. The process of determining whether or not the entered password
is correct is delegated to the server. If the password is correct, only
then are the changes made by the user are saved.