Made SideBar more consistent with Chatter-Layout proposal
This includes some re-writing of the SideBar component, and the introduction of the SideBarOption component. This was done to achieve a better separation of duties between SideBar and its child components. SideBarOption now handles the updating of the sessionStorage instance regarding the index of the currently selected SideBarOption (before this task was solely handled by SideBar). By doing this, SideBar is less saturated in terms of logic and has the sole purpose of being a display-only component.

Also includes the addition of icons and the logo used by the sidebar