Updated CommunicationEntity component to be consistent with Chatter-Layout
This includes the removal of the 'Action' version of a CommunicationEntity. This way of doing things seems inconsistent with most modern day web applications and could affect the user experience in a negative manner. Note that CommunicationEntityActionBar needs to be addressed/removed in a separate issue. This also decreases the complex logic to do with communication entities

This commit also rewrites some parts of communication-entity-styles.css for better readability and clearness.